Blade kajigger FreeCAD source and .stl files added
The blade kajigger is a tool for slitting long bits of felt or fabric in half, specifically 1/4"x1" polyester felt. The current design slides onto the side of my ikea table pretty nicely, which has a thickness of 27mm (1.06"). The freecad source is included so it shouldn't be difficult to modify for different tables/surfaces or felt strip sizes.